Fix scraping of "Spack Smoke Test" attribute
The Spack DSL for Smoke Tests recently changed, and the `test()`
method is now deprecated. Spack Smoke Tests now use a family of methods
named `test_*()`:

https://spack.readthedocs.io/en/latest/packaging_guide.html#adding-stand-alone-tests

As a result of this change, the DocPortal generator was incorrectly
reporting that many Spack packages lacked a Smoke Test when they in fact
have one or more.

This commit updates the scraping pattern in the DocPortal generator to
recognize either Smoke Test syntax, Also add a break statement to reduce
unnecessary churn.